requirements.txt
pdfplumber==0.11.7
pyyaml==6.0.2
flask==3.1.2
python-dateutil==2.9.0.post0
